Always choose alphabetically lesser value if tied
In case of a tie in which value is more common, we don't want the choice of which setting value to use to be non-deterministic based on the order of the dictionary. Instead, we choose the value that is earlier in alphabetical order. This also happens to choose formulas more often than plain values, which is nice.