Add test for whether non-material settings skip bubbling to materials
If they don't skip bubbling, the material profile might get the same value for the setting as the quality profile. The setting is then left out of the quality profile. The setting can't be saved in the XML file for the material, so the setting will be wrong in the quality profile.

Contributes to issue CURA-3650.